+++ docs/changes.rst
@@ -15,9 +15,12 @@ Version 0.5.4
 Unreleased.
 
  - Rewrote :attr:`~wand.api.libc` library loader. [:issue:`409`]
+ - Added :meth:`Image.brightness_contrast() <wand.image.BaseImage.brightness_contrast>` method.
  - Added :meth:`Image.oil_paint() <wand.image.BaseImage.oil_paint>` method.
  - Added :meth:`Image.opaque_paint() <wand.image.BaseImage.opaque_paint>` method.
  - Added :meth:`Image.rotational_blur() <wand.image.BaseImage.rotational_blur>` method.
+ - Added :meth:`Image.shear() <wand.image.BaseImage.shear>` method.
+ - Added :meth:`Image.sigmoidal_contrast() <wand.image.BaseImage.sigmoidal_contrast>` method.
 
 
 .. _changelog-0.5.3: